+++ src/equipment.cpp
@@ -2,54 +2,66 @@
 
 namespace Enchant {
 #pragma region Attribute
-int Attribute::callback(void *data, int argc, char **argv, char **column_name){
-    Attribute* self = (Attribute*) data;  // force to change type
-    self->_name = argv[0];
-    self->_text = argv[1];
+int            Attribute::callback(void* data, int argc, char** argv, char** column_name) {
+    Attribute* self = (Attribute*)data;  // force to change type
+    self->_name     = argv[0];
+    self->_text     = argv[1];
 
-    for(int i = 0; i < argc; i++){
+    for (int i = 0; i < argc; i++) {
         printf("%s = %s\n", column_name[i], argv[i] ? argv[i] : "NULL");
     }
     printf("\n");
     return 0;
 }
 
-Attribute::Attribute(int id, float value) : id(id), value(value) {
+Attribute::Attribute(int id, float value) : _id(id) {
     sqlite3* db;
     char*    err_msg = 0;
-    int      rc      = sqlite3_open("attribute.sqlite3", &db);
+    int      rc      = sqlite3_open(Attribute::db_name.c_str(), &db);
 
     // database don't be opened, then we close it and stop this function.
-    if (!rc){
+    if (rc) {
         fprintf(stderr, "[SQL error] Can't open database, %s\n", sqlite3_errmsg(db));
         sqlite3_close(db);
         return;
     }
 
     // create SQL statement to load data
-    std::string sql = "SELECT name, description FROM attribute WHERE number = " + std::to_string(id);
+    std::string sql =
+        "SELECT name, description FROM attribute WHERE number = " + std::to_string(id);
 
     // load data from database by executing SQL statement
     rc = sqlite3_exec(db, sql.c_str(), callback, (void*)this, &err_msg);
-    if( rc != SQLITE_OK ){
+    if (rc != SQLITE_OK) {
         fprintf(stderr, "[SQL error] %s\n", err_msg);
         sqlite3_free(err_msg);
     }
 
-    // test
-    this->substitute();
-    std::cout << "中文測試->" << this->_text << std::endl;
-
     sqlite3_close(db);
+
+    // NOTE: set value and format. It must be put behind loading data,
+    // otherwise, the data from database will cover the formatted text.
+    setValue(value);
+
+    // BUG: test
+    std::cout << "中文測試->" << this->getText() << std::endl;
 }
 
-void Attribute::substitute(){
-    _text.replace(_text.find('X'), 1, std::to_string(value));
-}
+// we must output correct value + text dynamicly, so copying _text and formatting it.
+std::string Attribute::getText() { return _text; }
 
 std::string Attribute::getName() { return _name; }
 
-std::string Attribute::getText() { return _text; }
+float Attribute::getValue() { return _value; }
+
+Attribute& Attribute::setValue(const float value) {
+    _value = value;
+
+    // replace X or old value with new value by regex
+    std::regex reg("\\d+.\\d+|X");
+    _text = std::regex_replace(_text, reg, std::to_string(value));
+    return *this;
+}
 #pragma endregion
 
 #pragma region AttributeTree