+++ src/main.cpp
@@ -3,8 +3,11 @@
 using namespace std;
 
 int main() {
+    std::regex reg("\\d+.\\d+|X");
+    cout << "regex: " << std::regex_replace("+X physical attack", reg, "4.0") << endl;
+    cout << "regex: " << std::regex_replace("+158.223156465 physical attack", reg, "4.0") << endl;
     Enchant::Attribute a(1), b(2, 5.);
-    cout << a.value << " " << b.value << endl;
+    cout << a.getValue() << " " << b.getValue() << endl;
 
     Enchant::AttributeTree c();